generic FontIconSet
