+++ QtAwesome/__init__.py
@@ -1,1 +1,7 @@
+"""
+PyQtAwesome - use font-awesome in PyQt / PySide applications
+
+This is a port to Python of the C++ QtAwesome by Rick Blommers
+"""
+
 from .fa import QtAwesome, codes, bunch
